[#3513] encoding app launch url before redirect
