+++ hs_tracking/views.py
@@ -11,7 +11,7 @@ from django.contrib import messages
 from . import models as hs_tracking
 from .models import Session, Variable
 from .utils import get_std_log_fields, authentic_redirect_url
-from hs_tools_resource.utils import do_work_when_launching_app_as_needed, WebAppLaunchException
+from hs_tools_resource.utils import do_work_when_launching_app_as_needed, WebAppLaunchException, split_url
 
 
 class AppLaunch(TemplateView):
@@ -59,6 +59,15 @@ class AppLaunch(TemplateView):
             # format and save the log message
             msg = Variable.format_kwargs(**fields)
             session.record('app_launch', value=msg)
+
+        # encode url before redirect
+        path, query = split_url(url)
+        path = urllib.parse.quote(path).replace('%3A', ':')
+        if query:
+            query = urllib.parse.quote(query, safe='')
+            url = "{}?{}".format(path, query)
+        else:
+            url = path
 
         return HttpResponseRedirect(url)
